# lib/api-spec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Valo Pay sandbox API",
    "version": "1.1.0",
    "description": "Valo Pay collections and connected banking sandbox API. All monetary fields are integer minor units (NGN kobo). Real data and all outbound provider instructions are disabled in connected modules."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/v1/workspace": {
      "get": {
        "operationId": "getWorkspace",
        "tags": [
          "valopay"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Permission or readiness gate blocked"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "summary": "The caller's workspace: its lenders, roles and actor",
        "description": "On a first visit an anonymous caller gets a new synthetic sandbox with two lenders; a signed-in person gets their own workspace. New sandboxes are limited per client address."
      }
    },
    "/v1/overview": {
      "get": {
        "operationId": "getOverview",
        "tags": [
          "valopay"
        ],
        "parameters": [
          {
            "name": "merchantId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The lender (a merchant in the API) the request is scoped to; one of the caller's workspace merchants."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Overview"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Permission or readiness gate blocked"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "summary": "The operations overview for one lender",
        "description": "Metrics, queues, recent activity, upcoming due items, the last and next daily close, and the alerts feed (NFR-OBS-02)."
      }
    },
    "/v1/records/{kind}": {
      "get": {
        "operationId": "listRecords",
        "tags": [
          "valopay"
        ],
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Record kind: one of the shared schema's recordKinds (customers, mandates, due-items, attempts, observations, payments, ...)."
          },
          {
            "name": "merchantId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The lender (a merchant in the API) the request is scoped to; one of the caller's workspace merchants."
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Text matched, ignoring case and accents, against the name, reference, status and data."
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Only records in this status; omitted or \"all\" for every status."
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500
            },
            "description": "Page size, capped at 500 when supplied. Omitted returns the complete filtered kind for existing relationship and balance views."
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            },
            "description": "Rows to skip in the newest-first order."
          },
          {
            "name": "updatedSince",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "ISO timestamp; only records updated at or after it (incremental sync)."
          },
          {
            "name": "customerId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Only records directly linked to this customer, in the selected lender."
          },
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Only this exact record ID, in the selected kind and lender."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordList"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Permission or readiness gate blocked"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "summary": "Records of one kind for one lender, newest first",
        "description": "Filtered by status and by a search that ignores case and accents; paged with limit and offset; updatedSince for incremental sync."
      },
      "post": {
        "operationId": "createRecord",
        "tags": [
          "valopay"
        ],
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Record kind: one of the shared schema's recordKinds (customers, mandates, due-items, attempts, observations, payments, ...)."
          },
          {
            "name": "merchantId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The lender (a merchant in the API) the request is scoped to; one of the caller's workspace merchants."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValopayRecord"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Permission or readiness gate blocked"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordInput"
              }
            }
          }
        },
        "summary": "Create a record of an editable kind",
        "description": "Validated against the kind's data schema; a status only a domain action may set is refused."
      }
    },
    "/v1/records/{kind}/{id}": {
      "patch": {
        "operationId": "updateRecord",
        "tags": [
          "valopay"
        ],
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Record kind: one of the shared schema's recordKinds (customers, mandates, due-items, attempts, observations, payments, ...)."
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The record's id."
          },
          {
            "name": "merchantId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The lender (a merchant in the API) the request is scoped to; one of the caller's workspace merchants."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValopayRecord"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Permission or readiness gate blocked"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordUpdate"
              }
            }
          }
        },
        "summary": "Update a record",
        "description": "Editable kinds only; an approved, preregistered or closed version is immutable. Send expectedUpdatedAt from the edit's original record to reject stale changes with 409. An identical successful Idempotency-Key replay returns its original result before checking the version."
      }
    },
    "/v1/actions": {
      "post": {
        "operationId": "performAction",
        "tags": [
          "valopay"
        ],
        "parameters": [
          {
            "name": "merchantId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The lender (a merchant in the API) the request is scoped to; one of the caller's workspace merchants."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Permission or readiness gate blocked"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionInput"
              }
            }
          }
        },
        "summary": "Run a domain action on the lender's state",
        "description": "Every action is audited, most require a reason, and the persona's role applies; the catalogue of actions is in docs/frontend-contract.md."
      }
    },
    "/v1/imports": {
      "post": {
        "operationId": "importRecords",
        "tags": [
          "valopay"
        ],
        "parameters": [
          {
            "name": "merchantId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The lender (a merchant in the API) the request is scoped to; one of the caller's workspace merchants."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Permission or readiness gate blocked"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportInput"
              }
            }
          }
        },
        "summary": "Preview or commit a synthetic CSV import",
        "description": "commit=false validates every row and reports each; commit=true persists all rows or none. syntheticOnly must be true: no real lender data."
      }
    },
    "/v1/customers/{id}/timeline": {
      "get": {
        "operationId": "getCustomerTimeline",
        "tags": [
          "valopay"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The record's id."
          },
          {
            "name": "merchantId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The lender (a merchant in the API) the request is scoped to; one of the caller's workspace merchants."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Timeline"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Permission or readiness gate blocked"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "summary": "A customer's position and complete timeline",
        "description": "Every event, mandate, due item and payment, with each retry decision as it was recorded."
      }
    },
...
    "/v1/gates": {
      "get": {
        "operationId": "getGates",
        "tags": [
          "valopay"
        ],
        "parameters": [
          {
            "name": "merchantId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The lender (a merchant in the API) the request is scoped to; one of the caller's workspace merchants."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Gates"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Permission or readiness gate blocked"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "summary": "Production readiness gates",
        "description": "Prerequisites and decisions, always unproven on synthetic data, and the limitations the sandbox cannot remove."
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Workspace": {
        "type": "object",
        "properties": {
          "accessMode": { "type": "string", "enum": ["sandbox", "staff"], "description": "Whether the server authorises a demo persona or a provisioned staff membership." },
          "viewerScope": { "type": "string", "description": "Opaque workspace/user scope for browser preferences; never an authorisation credential." },
          "name": {
            "type": "string"
          },
          "environment": {
            "type": "string"
          },
          "actor": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "authenticated": {
            "type": "boolean"
          },
          "merchants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Merchant"
            }
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "productionEnabled": {
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "environment",
          "actor",
          "role",
          "authenticated",
          "merchants",
          "roles",
          "productionEnabled"
        ],
        "description": "The caller's workspace: who is acting, in which role, whether they signed in, and the lenders and roles available."
      },
...
    }
  }
}